Generate reduce actions
If a reduce/shift conflict is detected, a warning is produced and the
reduce is prioritized. Finer grain control of this needs to be added.
If a reduce/reduce conflict is detected, a warning is produced and one
of the reductions is picked. Control of this needs to be added.
Reduce actions are produced on each terminal.