Allow empty production
Change the production grammar so that there doesn't need to be an
ident between the '=>' and the ';'. Every place else is okay with an
empty prodcution. It simply wasn't allowed by the grammar.

disable the google-runtime-references check in clangtidy. The idea that
a paramater must be a const& or * (never a mutable reference) to be a
bit silly. I'm not sure what that is protecting against.