Add better control for class names
Split things so that there are three different controls for names.
- "lexer class X" for the lexer name
- "parser class X" for the parser name
- "namespace X" for the enclosing namespace.

"parser class" no long controls the name of the output file. That is
only controlled by the -o option. The default name of the output file is
now based on the input file name rather than the class.

Fix some syntax issues with the sqlite grammar. This file still needs a
lot of love and attention.