Add semantic actions for rules
It is now possible to assign semantic actions to each of the productions
in a rule. The semanticvalues for the items in the production are in
variables of the form `_v{n}` where `{n}` is the position of the item
in the production starting from the right, numberred starting at one.

items may also be assigned aliases. Those aliases will be refernce
symonyms for the corresponding `_v{n}`.

Example:

rule <int> add {
     => left:INT '+' right:INT <%{ return left + right; }%>
}

or, without the aliases:

rule <int> add { => INT '+' INT <%{ return _v1 + _v3; }%> }