Allow inline names in the namespace option
Enhancement #16

The namespace option now can take a normal identifier (for simple
namespaces) or a single-quotes string for inline name spaces.
```
namespace 'one::two';
```

As a part of this, I updated the test runner so that it coudld actually
run yalr and examine the generated code. This allowed me to add explicit
tests for the `namespace`, `lexer class`, and `parser class` options.