Add associativity and precedence
You can now specify associativity and precedence for terms and
precedence for productions. These are used to resolve shift/reduce and
reduce/reduce conflicts in the grammar.

This completes issue #20

This commit also resolves issue #15 - moving the analyzer to use
error_info. More work needs to be done (globally) to use the nested
structure of error_info to its advantage, but this at least makes the
analyzer use the infrastructure.