Fix up examples so they at least generate.
Many of the examples used older syntax or had things that wouldn't
parse. This commit fixes them so that they at least parse and generate
code. No work was done to make sure the grammars were in any sense
"correct".

In particular, the sqlite grammar has some known wrongness baked in.
These are marked in the grammar.

Along the way, some bugs in the parser and error reporting were also
fixed.