Add "global" flag to terminals
in support of enhancement #34
Add the ability to mark a terminal as "global":
global term $type $name ...

Make sure the globl flag persists in the symbol table through to
codegen. Update the generated code to incude the flag. Note, it
currently doesn't do anything with it, but it is there.