fix reduce of epsilon productions
There was a bug in the new table driven parser when reducing epsilon
productions. The new code tried to get the state from the lowest most
used stack item. But for an epislon reduction, there isn't one. So, use
the current state as the state to use when consulting the goto table.