refactor slr parser generator
Refactor the SLR parser generator to make a class out of it. This will
add the necessary hooks to allow multiple backends in preparation for
issue #38 and #11.

All backends are now subclasses of yalr::algo::parser_generator.

By convention source code will be in lib/algo and include/algo prefixed
by the name of the algorithm.