start refactoring code generation
AS a prelude for issues #11 and #38, refactor the code generation. All
three algorithms will share a common lexer. So add a driver layer that
does all the preamble and the lexer and (possibly) the embedded main.
That wat the parser_generator object is only resonsible for the actual
parser output.

For now - just add the wrapper layer and move the main generation to
that. Leave the lexer for the next step.