Fix long standing bug in confilct resolution
The code that check precedence and associativty to (possibly) resolve
conflicts has a bug where it assumed that any conflict that involved a
terminal as the symbol must be a shift/reduce conflict. But that is not
true. Instead, actually check the current incumbent action to see if it
is a reduce or not to decide if it is a shift/redecude or reduce/reduce
conflict.